requirements.txt
# python packages needed
openai
anthropic
google-generativeai

firestore
black
tiktoken
chromadb
transformers
sentence-transformers
bs4
firecrawl-py
streamlit
pypdf2
python-dotenv
faiss-cpu
huggingface_hub
psycopg2
beautifulsoup4
yfinance
uvicorn
rich
tavily-python
grandalf

# langchain common modules
langchain
langchain-community
langchain-tavily
langchainhub
wikipedia

# for OpenAI
langchain-openai
# for accessing Google Gemini
langchain-google-genai
langchain-google-firestore
# for Anthropic Claude
langchain-anthropic
# for Groq
langchain-groq
# for huggingface
langchain-huggingface





